% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.r
\name{delta}
\alias{delta}
\title{delta}
\usage{
delta(rt, compatible, sujet = NULL, quant = c(1:10)/10)
}
\arguments{
\item{rt}{Response time}

\item{compatible}{compatible vs incompatible trials, 'c' for compatible trials, 'i' for incompatible}

\item{sujet}{subject names. Default is NULL}

\item{quant}{quantiles}
}
\description{
This function compute classical delta plots
}
\examples{
delta(df$rt,  df$compatible, df$sujet, quant = c(1:5)/5) 
Return L(p)  proportion of errors among the pth first quantile of response time (computed on all responses)
}
\keyword{dela}
\keyword{plot}
