% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elf.r
\name{elf.g}
\alias{elf.g}
\title{ELF group}
\usage{
elf.g(RT, accuracy, sujet, na.rm = FALSE)
}
\arguments{
\item{RT}{response time}

\item{accuracy}{accuracy, 1 for correct trials, 0 for errors. Ex: df$accuracy}

\item{sujet}{name of subjects Ex: df$sujet}
}
\description{
This function compute aggregated  Error Location Index for groups
}
\examples{
elf.g(df$RT, df$accuracy, df$sujet) 
}
\keyword{Error}
\keyword{Index}
\keyword{Location}
