% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elf.r
\name{eli}
\alias{eli}
\title{ELI}
\usage{
eli(L, alpha = 1, na.rm = FALSE)
}
\arguments{
\item{L}{an ELF function}

\item{alpha}{weight of ELI function. Default is 1. The greater alpha, the more weight is given to fast responses}
}
\description{
This function compute Error Location Index
}
\examples{
eli(L) 
}
\keyword{Error}
\keyword{Index}
\keyword{Location}
