% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.r
\name{lorenz}
\alias{lorenz}
\title{Lorenz-Delta}
\usage{
lorenz(rt, comp)
}
\arguments{
\item{rt}{Response time}

\item{comp}{compatible vs incompatible trials, 'c' for compatible trials, 'i' for incompatible}
}
\description{
This function compute  Lorenz-Dela plot
}
\examples{
lorenz(rt, comp)
Return  inhibition index and Lorenz-Delta plot
}
\keyword{Lorenz-Delta}
\keyword{plots}
