% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vincentization.r
\name{vincentize}
\alias{vincentize}
\title{Vincentization}
\usage{
vincentize(dataframe, rt, accuracy = NA, var, quantiles)
}
\arguments{
\item{dataframe}{the data frame  df containing data}

\item{rt}{response time, ex: df$rt}

\item{accuracy}{accuracy, 1 for correct trials, 0 for errors. Ex: df$accuracy}

\item{var}{list variables within which vincentization will be performed}

\item{quantiles}{vector of quantiles used for vincentization}
}
\description{
This function vincentize da data frame
}
\examples{
 vincentize(data,data$rt,data$acc,list(sujet=data$sujet, compatible=data$compatible),c(0.2,0.4,0.6,0.8))
compute the means for quantiles (0-0.2,0.2-0.4,0.4-0.6,0.6-0.8,0.8-1)
for the variables data$rt and data$acc 
for each Sujet, in each modalities of compatible
return new variables:
 q = quantile
nq = proportion of trials within quantile
rt_m = means of response time by quantile
accuracy_m = means of accuracy by quantile
}
\keyword{vincentization}
