% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elf.r
\name{elf}
\alias{elf}
\title{ELF}
\usage{
elf(rt, accuracy, na.rm = TRUE)
}
\arguments{
\item{rt}{Response time}

\item{accuracy}{accuracy, 1 for correct trials, 0 for errors. Ex: df$accuracy}

\item{na.rm}{should missing data be disregarded (if there is no error). Default is TRUE}
}
\description{
This function compute Error Location Function
}
\examples{
elf(df$rt, df$accuracy) 
Return L(p)  proportion of errors among the pth first quantile of response time (computed on all responses)
}
\keyword{Error}
\keyword{Function}
\keyword{Location}
