% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.r
\name{inhib.delta}
\alias{inhib.delta}
\title{inhibibition indices}
\usage{
inhib.delta(rt, comp, sujet = sujet, cond = NA, dquantile = 20)
}
\arguments{
\item{dquantile}{number of quantiles used to compute the slope of last segment}

\item{data}{dataframe with rt, comp,  sujet, cond}
}
\description{
This function compute inhibition indices
}
\examples{
inhib.delta(rt, comp, cond, dquantile = 20) 
rt: response time
comp: compatible vs incompatible ('c' vs 'i')
sujet: subjects id (transformed into factor if needed)
cond: experimental conditions (transformed into factor if needed)
dquantile: number of quantiles to compute last segment slope and linear approximation
Output: last segment delta slope, delta slopes, linear approximation (trend = coefficient, intercept), inhibition index
q, delta.slope, index
}
\keyword{dela}
\keyword{index}
\keyword{inhibition}
\keyword{plot}
